refactor: isolate express compatibility layer
